Update applicable course formats
Previously, this block was only usable with the Topics or Weekly
course formats. However, it functions perfectly well with the
Flexible Sections and Collapsed Topics formats, also, and so those
are now whitelisted for use with this block. I would have preferred
using a blacklist rather than a whitelist, but the Moodle code that
checks this doesn't check for true/false. Instead, it simply assumes
any entries present are able to use this block, and so we are stuck
with a whitelist (for now).