More flexible course format restriction logic
Instead of using an explicit whitelist of course formats allowed to
use this block, change applicable_formats() to allow any course format
to use this block while still proscribing its use for plugins and tags.
This is needed because there are many course formats, some of which
are custom and, thus, specific to a single organization (and yet should
still be able to use this block).