Improve checkbox processing
A problem was discovered where the checkbox variable would be null
when a non-input HTML element, such as LI, had an id conforming to
the pattern expected by Mass Actions when attempting to carry out
the user's selected action. This resulted in a Javascript error
that rendered the block unusable and made the page appear
unresponsive to the user. By simply checking that the checkbox
variable is not null first, we can ensure we skip and do not attempt
to process non-input elements.